// src/generated/graphql.ts
export type Maybe<T> = T | null
// ...
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  JSON: any
}
// ...
export type AppointmentModel = {
  __typename?: 'AppointmentModel'
  _eTag?: Maybe<Scalars['String']>
  _embedded?: Maybe<EmbeddedAppointmentModel>
  _links?: Maybe<Scalars['JSON']>
  accompanied?: Maybe<Scalars['Boolean']>
  appointmentType?: Maybe<ListItemModel>
  attendee?: Maybe<AppointmentAttendeeModel>
  attendeeConfirmed?: Maybe<Scalars['Boolean']>
  cancelled?: Maybe<Scalars['Boolean']>
  created?: Maybe<Scalars['String']>
  description?: Maybe<Scalars['String']>
  end?: Maybe<Scalars['String']>
  followUp?: Maybe<AppointmentModelFollowUp>
  id?: Maybe<Scalars['String']>
  metadata?: Maybe<Scalars['JSON']>
  modified?: Maybe<Scalars['String']>
  negotiatorConfirmed?: Maybe<Scalars['Boolean']>
  negotiatorIds?: Maybe<Array<Scalars['String']>>
  negotiators?: Maybe<Array<Maybe<NegotiatorModel>>>
  officeIds?: Maybe<Array<Scalars['String']>>
  offices?: Maybe<Array<Maybe<OfficeModel>>>
  organiserId?: Maybe<Scalars['String']>
  property?: Maybe<PropertyModel>
  propertyConfirmed?: Maybe<Scalars['Boolean']>
  propertyId?: Maybe<Scalars['String']>
  recurring?: Maybe<Scalars['Boolean']>
  start?: Maybe<Scalars['String']>
  typeId?: Maybe<Scalars['String']>
}
// ...
export type AreaModel = {
  __typename?: 'AreaModel'
  _eTag?: Maybe<Scalars['String']>
  _embedded?: Maybe<Scalars['JSON']>
  _links?: Maybe<Scalars['JSON']>
  active?: Maybe<Scalars['Boolean']>
  area?: Maybe<Array<Scalars['String']>>
  created?: Maybe<Scalars['String']>
  departmentIds?: Maybe<Array<Scalars['String']>>
  id?: Maybe<Scalars['String']>
  modified?: Maybe<Scalars['String']>
  name?: Maybe<Scalars['String']>
  officeIds?: Maybe<Array<Scalars['String']>>
  type?: Maybe<Scalars['String']>
}
// ...
export type DepartmentModel = {
  __typename?: 'DepartmentModel'
  _eTag?: Maybe<Scalars['String']>
  _embedded?: Maybe<Scalars['JSON']>
  _links?: Maybe<Scalars['JSON']>
  ageOptions?: Maybe<Array<Scalars['String']>>
  created?: Maybe<Scalars['String']>
  id?: Maybe<Scalars['String']>
  localityOptions?: Maybe<Array<Scalars['String']>>
  modified?: Maybe<Scalars['String']>
  name?: Maybe<Scalars['String']>
  parkingOptions?: Maybe<Array<Scalars['String']>>
  situationOptions?: Maybe<Array<Scalars['String']>>
  styleOptions?: Maybe<Array<Scalars['String']>>
  typeOptions?: Maybe<Array<Scalars['String']>>
}

export type DepartmentModelPagedResult = {
  __typename?: 'DepartmentModelPagedResult'
  _embedded?: Maybe<Array<DepartmentModel>>
  _links?: Maybe<Scalars['JSON']>
  pageCount?: Maybe<Scalars['Int']>
  pageNumber?: Maybe<Scalars['Int']>
  pageSize?: Maybe<Scalars['Int']>
  totalCount?: Maybe<Scalars['Int']>
  totalPageCount?: Maybe<Scalars['Int']>
}

export type DocumentModel = {
  __typename?: 'DocumentModel'
  _eTag?: Maybe<Scalars['JSON']>
  _embedded?: Maybe<EmbeddedDocumentModel>
  _links?: Maybe<Scalars['JSON']>
  associatedId?: Maybe<Scalars['String']>
  associatedType?: Maybe<Scalars['String']>
  created?: Maybe<Scalars['String']>
  id?: Maybe<Scalars['String']>
  modified?: Maybe<Scalars['String']>
  name?: Maybe<Scalars['String']>
  typeId?: Maybe<Scalars['String']>
}
// ...
export type EmbeddedPropertyModel = {
  __typename?: 'EmbeddedPropertyModel'
  appointments?: Maybe<Array<Maybe<AppointmentModel>>>
  area?: Maybe<AreaModel>
  department?: Maybe<DepartmentModel>
  documents?: Maybe<Array<Maybe<DocumentModel>>>
  images?: Maybe<Array<Maybe<PropertyImageModel>>>
  landlord?: Maybe<LandlordModel>
  negotiator?: Maybe<NegotiatorModel>
  offers?: Maybe<Array<Maybe<OfferModel>>>
  tenancies?: Maybe<Array<Maybe<TenancyModel>>>
  vendor?: Maybe<VendorModel>
}
// ...
export type KeyModel = {
  __typename?: 'KeyModel'
  created?: Maybe<Scalars['String']>
  id: Scalars['ID']
  keysInSet?: Maybe<Array<Maybe<IndividualKeyModel>>>
  modified?: Maybe<Scalars['String']>
  movements?: Maybe<Array<Maybe<KeyMovementModel>>>
  number?: Maybe<Scalars['String']>
  office?: Maybe<OfficeModel>
  status?: Maybe<KeyStatus>
  type?: Maybe<KeyType>
}
// ...
export enum KeyStatus {
  CheckedIn = 'checkedIn',
  CheckedOut = 'checkedOut',
  NoLongerHeld = 'noLongerHeld',
}

export type KeyType = {
  __typename?: 'KeyType'
  id?: Maybe<Scalars['String']>
  value?: Maybe<Scalars['String']>
}

export type LandLordSourceInput = {
  id?: InputMaybe<Scalars['String']>
  type?: InputMaybe<Scalars['String']>
}

export type LandlordModel = {
  __typename?: 'LandlordModel'
  _eTag?: Maybe<Scalars['String']>
  _embedded?: Maybe<EmbeddedLandlordModel>
  _links?: Maybe<Scalars['JSON']>
  active?: Maybe<Scalars['Boolean']>
  created?: Maybe<Scalars['String']>
  id?: Maybe<Scalars['String']>
  metadata?: Maybe<Scalars['JSON']>
  modified?: Maybe<Scalars['String']>
  officeId?: Maybe<Scalars['String']>
  related?: Maybe<Array<LandlordRelatedModel>>
  solicitorId?: Maybe<Scalars['String']>
  source?: Maybe<LandlordSourceModel>
}
// ...
export type NegotiatorModel = {
  __typename?: 'NegotiatorModel'
  _eTag?: Maybe<Scalars['String']>
  _embedded?: Maybe<EmbeddedNegotiatorModel>
  _links?: Maybe<Scalars['JSON']>
  active?: Maybe<Scalars['Boolean']>
  created?: Maybe<Scalars['String']>
  email?: Maybe<Scalars['String']>
  id?: Maybe<Scalars['String']>
  jobTitle?: Maybe<Scalars['String']>
  metadata?: Maybe<Scalars['JSON']>
  mobilePhone?: Maybe<Scalars['String']>
  modified?: Maybe<Scalars['String']>
  name?: Maybe<Scalars['String']>
  officeId?: Maybe<Scalars['String']>
  workPhone?: Maybe<Scalars['String']>
}
// ...
export type OfferModel = {
  __typename?: 'OfferModel'
  OfferId?: Maybe<Scalars['String']>
  _eTag?: Maybe<Scalars['String']>
  _embedded?: Maybe<EmbeddedOfferModel>
  _links?: Maybe<Scalars['JSON']>
  amount?: Maybe<Scalars['String']>
  applicantId?: Maybe<Scalars['String']>
  conditions?: Maybe<Scalars['String']>
  created?: Maybe<Scalars['String']>
  currency?: Maybe<Scalars['String']>
  date?: Maybe<Scalars['String']>
  exclusions?: Maybe<Scalars['String']>
  id?: Maybe<Scalars['String']>
  inclusions?: Maybe<Scalars['String']>
  metadata?: Maybe<Scalars['JSON']>
  modified?: Maybe<Scalars['String']>
  propertyId?: Maybe<Scalars['String']>
  related?: Maybe<Array<Maybe<OfferRelate>>>
  status?: Maybe<Scalars['String']>
}
// ...
export type PropertyImageModel = {
  __typename?: 'PropertyImageModel'
  _eTag?: Maybe<Scalars['String']>
  _embedded?: Maybe<EmbeddedPropertyImageModel>
  _links?: Maybe<Scalars['JSON']>
  caption?: Maybe<Scalars['String']>
  created?: Maybe<Scalars['String']>
  id?: Maybe<Scalars['String']>
  modified?: Maybe<Scalars['String']>
  order?: Maybe<Scalars['Int']>
  propertyId?: Maybe<Scalars['String']>
  type?: Maybe<Scalars['String']>
  url?: Maybe<Scalars['String']>
}

export enum PropertyImageQueryEmbed {
  Property = 'property',
}

export type PropertyModeLetting = {
  __typename?: 'PropertyModeLetting'
  agentRole?: Maybe<Scalars['String']>
  availableFrom?: Maybe<Scalars['String']>
  availableTo?: Maybe<Scalars['String']>
  brochureId?: Maybe<Scalars['String']>
  furnishing?: Maybe<Array<Scalars['String']>>
  instructed?: Maybe<Scalars['String']>
  landlordId?: Maybe<Scalars['String']>
  lettingFee?: Maybe<PropertyModelLettingFee>
  managementFee?: Maybe<PropertyModelManagementFee>
  rent?: Maybe<Scalars['Float']>
  rentFrequency?: Maybe<Scalars['String']>
  status?: Maybe<Scalars['String']>
  term?: Maybe<Scalars['String']>
}

export type PropertyModel = {
  __typename?: 'PropertyModel'
  _eTag?: Maybe<Scalars['String']>
  _embedded?: Maybe<EmbeddedPropertyModel>
  _links?: Maybe<Scalars['JSON']>
  address?: Maybe<PropertyModelAddress>
  age?: Maybe<Array<Scalars['String']>>
  alternateId?: Maybe<Scalars['String']>
  archivedOn?: Maybe<Scalars['String']>
  areaId?: Maybe<Scalars['String']>
  bathrooms?: Maybe<Scalars['Int']>
  bedrooms?: Maybe<Scalars['Int']>
  councilTax?: Maybe<Scalars['String']>
  created?: Maybe<Scalars['String']>
  currency?: Maybe<Scalars['String']>
  departmentId?: Maybe<Scalars['String']>
  description?: Maybe<Scalars['String']>
  epc?: Maybe<PropertyModelEpc>
  externalArea?: Maybe<PropertyModelExternalArea>
  fromArchive?: Maybe<Scalars['Boolean']>
  id?: Maybe<Scalars['String']>
  internalArea?: Maybe<PropertyModelInternalArea>
  internetAdvertising?: Maybe<Scalars['Boolean']>
  keys?: Maybe<Array<Maybe<KeyModel>>>
  letting?: Maybe<PropertyModeLetting>
  locality?: Maybe<Array<Scalars['String']>>
  longDescription?: Maybe<Scalars['String']>
  lostInstructionDate?: Maybe<Scalars['String']>
  lostInstructionNote?: Maybe<Scalars['String']>
  marketingMode?: Maybe<Scalars['String']>
  metadata?: Maybe<Scalars['JSON']>
  modified?: Maybe<Scalars['String']>
  negotiatorId?: Maybe<Scalars['String']>
  notes?: Maybe<Scalars['String']>
  officeIds?: Maybe<Array<Scalars['String']>>
  parking?: Maybe<Array<Scalars['String']>>
  receptions?: Maybe<Scalars['Int']>
  rooms?: Maybe<Array<PropertyRoomModel>>
  selling?: Maybe<PropertyModelSelling>
  situation?: Maybe<Array<Scalars['String']>>
  strapline?: Maybe<Scalars['String']>
  style?: Maybe<Array<Scalars['String']>>
  summary?: Maybe<Scalars['String']>
  type?: Maybe<Array<Scalars['String']>>
  video2Caption?: Maybe<Scalars['String']>
  video2Url?: Maybe<Scalars['String']>
  videoCaption?: Maybe<Scalars['String']>
  videoUrl?: Maybe<Scalars['String']>
  viewingArrangements?: Maybe<Scalars['String']>
}

export type PropertyModelAddress = {
  __typename?: 'PropertyModelAddress'
  buildingName?: Maybe<Scalars['String']>
  buildingNumber?: Maybe<Scalars['String']>
  countryId?: Maybe<Scalars['String']>
  geolocation?: Maybe<PropertyModelAddressGeolocation>
  line1?: Maybe<Scalars['String']>
  line2?: Maybe<Scalars['String']>
  line3?: Maybe<Scalars['String']>
  line4?: Maybe<Scalars['String']>
  postcode?: Maybe<Scalars['String']>
}

export type PropertyModelAddressGeolocation = {
  __typename?: 'PropertyModelAddressGeolocation'
  latitude?: Maybe<Scalars['Float']>
  longitude?: Maybe<Scalars['Float']>
}

export type PropertyModelEpc = {
  __typename?: 'PropertyModelEpc'
  eer?: Maybe<Scalars['Int']>
  eerPotential?: Maybe<Scalars['Int']>
  eir?: Maybe<Scalars['Int']>
  eirPotential?: Maybe<Scalars['Int']>
  exempt?: Maybe<Scalars['Boolean']>
}

export type PropertyModelExternalArea = {
  __typename?: 'PropertyModelExternalArea'
  max?: Maybe<Scalars['Float']>
  min?: Maybe<Scalars['Float']>
  type?: Maybe<Scalars['String']>
}

export type PropertyModelFee = {
  __typename?: 'PropertyModelFee'
  amount?: Maybe<Scalars['Int']>
  type?: Maybe<Scalars['String']>
}

export type PropertyModelInternalArea = {
  __typename?: 'PropertyModelInternalArea'
  max?: Maybe<Scalars['Float']>
  min?: Maybe<Scalars['Float']>
  type?: Maybe<Scalars['String']>
}

export type PropertyModelLettingFee = {
  __typename?: 'PropertyModelLettingFee'
  amount?: Maybe<Scalars['Float']>
  type?: Maybe<Scalars['String']>
}

export type PropertyModelManagementFee = {
  __typename?: 'PropertyModelManagementFee'
  amount?: Maybe<Scalars['Float']>
  type?: Maybe<Scalars['String']>
}

export type PropertyModelSelling = {
  __typename?: 'PropertyModelSelling'
  agency?: Maybe<Scalars['String']>
  brochureId?: Maybe<Scalars['String']>
  completed?: Maybe<Scalars['String']>
  disposal?: Maybe<Scalars['String']>
  exchanged?: Maybe<Scalars['String']>
  fee?: Maybe<PropertyModelFee>
  instructed?: Maybe<Scalars['String']>
  price?: Maybe<Scalars['Float']>
  qualifier?: Maybe<Scalars['String']>
  recommendedPrice?: Maybe<Scalars['Int']>
  status?: Maybe<Scalars['String']>
  tenure?: Maybe<PropertyModelSellingTenure>
  vendorId?: Maybe<Scalars['String']>
}

export type PropertyModelSellingTenure = {
  __typename?: 'PropertyModelSellingTenure'
  expiry?: Maybe<Scalars['String']>
  type?: Maybe<Scalars['String']>
}

export enum PropertyQueryEmbed {
  Appointments = 'appointments',
  Area = 'area',
  Department = 'department',
  Documents = 'documents',
  Images = 'images',
  Landlord = 'landlord',
  Negotiator = 'negotiator',
  Offers = 'offers',
  Tenancies = 'tenancies',
  Vendor = 'vendor',
}

export type PropertyRoomModel = {
  __typename?: 'PropertyRoomModel'
  description?: Maybe<Scalars['String']>
  dimensions?: Maybe<Scalars['String']>
  name?: Maybe<Scalars['String']>
}
// ...
export type VendorModel = {
  __typename?: 'VendorModel'
  _eTag?: Maybe<Scalars['JSON']>
  _embedded?: Maybe<EmbeddedVendorModel>
  _links?: Maybe<Scalars['JSON']>
  archivedOn?: Maybe<Scalars['String']>
  created?: Maybe<Scalars['String']>
  fromArchive?: Maybe<Scalars['Boolean']>
  id?: Maybe<Scalars['String']>
  lastCall?: Maybe<Scalars['String']>
  metadata?: Maybe<Scalars['JSON']>
  modified?: Maybe<Scalars['String']>
  negotiatorId?: Maybe<Scalars['String']>
  nextCall?: Maybe<Scalars['String']>
  officeIds?: Maybe<Array<Maybe<Scalars['String']>>>
  propertyId?: Maybe<Scalars['String']>
  related?: Maybe<Array<Maybe<VendorRelate>>>
  sellingReasonId?: Maybe<Scalars['String']>
  solicitorId?: Maybe<Scalars['String']>
  source?: Maybe<VendorSourceModel>
  typeId?: Maybe<Scalars['String']>
}